Reorganize types package structure to make request type the root.
