+++ pkg/types/change/network/types.proto
@@ -16,12 +16,12 @@ limitations under the License.
 
 syntax = "proto3";
 
-package onos.config.network.change;
+package onos.config.change.network;
 
 import "google/protobuf/timestamp.proto";
 import "gogoproto/gogo.proto";
-import "github.com/onosproject/onos-config/pkg/types/meta/types.proto";
-import "github.com/onosproject/onos-config/pkg/types/device/change/types.proto";
+import "github.com/onosproject/onos-config/pkg/types/change/types.proto";
+import "github.com/onosproject/onos-config/pkg/types/change/device/types.proto";
 
 // NetworkChange specifies the configuration for a network change
 // A network change is a configuration change that spans multiple devices. The change contains a list of
@@ -43,10 +43,10 @@ message NetworkChange {
     uint64 revision = 3 [(gogoproto.casttype) = "Revision"];
 
     // 'status' is the current lifecycle status of the change
-    onos.config.meta.Status status = 4;
+    onos.config.change.Status status = 4;
 
     // 'reason' indicates the reason the change failed if in the FAILED state
-    onos.config.meta.FailureReason reason = 5;
+    onos.config.change.FailureReason reason = 5;
 
     // 'message' is an optional status message
     string message = 6;
@@ -59,5 +59,5 @@ message NetworkChange {
 
     // 'changes' is a set of changes to apply to devices
     // The list of changes should contain only a single change per device/version pair.
-    repeated onos.config.device.change.Change changes = 9;
+    repeated onos.config.change.device.Change changes = 9;
 }
